+++ packages/assets/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @ant-design/web3-assets
 
+## 1.14.3
+
+### Patch Changes
+
+- Updated dependencies [dfa67c5]
+- Updated dependencies [2ae483c]
+  - @ant-design/web3-common@1.22.0
+
+## 1.14.2
+
+### Patch Changes
+
+- Updated dependencies [cf5810d]
+  - @ant-design/web3-icons@1.16.0
+
 ## 1.14.1
 
 ### Patch Changes