+++ packages/wagmi/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @ant-design/web3-wagmi
 
+## 2.13.0
+
+### Minor Changes
+
+- 2ae483c: feat: support ignoreConfig for support multiple chain better
+
+### Patch Changes
+
+- Updated dependencies [dfa67c5]
+- Updated dependencies [2ae483c]
+  - @ant-design/web3-common@1.22.0
+  - @ant-design/web3-assets@1.14.3
+
+## 2.12.2
+
+### Patch Changes
+
+- @ant-design/web3-assets@1.14.2
+
 ## 2.12.1
 
 ### Patch Changes