+++ packages/web3/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @ant-design/web3
 
+## 1.26.0
+
+### Minor Changes
+
+- 2ae483c: feat: support ignoreConfig for support multiple chain better
+
+### Patch Changes
+
+- Updated dependencies [dfa67c5]
+- Updated dependencies [2ae483c]
+  - @ant-design/web3-common@1.22.0
+  - @ant-design/web3-assets@1.14.3
+
+## 1.25.2
+
+### Patch Changes
+
+- Updated dependencies [cf5810d]
+  - @ant-design/web3-icons@1.16.0
+  - @ant-design/web3-assets@1.14.2
+
 ## 1.25.1
 
 ### Patch Changes